requirements.txt
# Core dependencies
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-google-genai>=1.0.0
google-generativeai>=0.7.0

# Web interface
streamlit==1.39.0
streamlit-chat==0.1.1

# Data handling
pandas==2.2.2
python-dateutil==2.9.0

# Utilities
python-dotenv==1.0.1
